# FSharp.Core: Set: optimize tree layout