@forki 2020-09-27
Can you please put the map fixes into a separate PR. I assume it should be merged asap