@buybackoff 2020-09-27
Does anyone have an idea where to starts with the errors? All `FSharp.Core` tests are passing. Locally I have different errors vs CI, and CI output is completely useless.

I have a candidate for review - a monstrous [method `compareStacks`](https://github.com/dotnet/fsharp/pull/10190/files#diff-11ea2797399140c305791bf3e8e23d8dR467-R519), rewriting which took more time than everything else. But it failed initially in tests and then I fixed it. Maybe test coverage misses something, will review again.